Re-localized general functions to rabies.utils
